CRITICAL FIX: Correct RTP payload byte order to big-endian
BUG: Wrong byte order causing noise in all audio and corrupted IQ data
=====================================================================

ROOT CAUSE:
-----------
RTP payloads use NETWORK BYTE ORDER (big-endian) for all data.
We were parsing int16 samples as little-endian (numpy default).

This caused:
- Audio streaming: complete noise (random byte swapping)
- GRAPE IQ data: corrupted complex samples
- All doppler measurements: INVALID

EVIDENCE:
---------
Testing with SSRC 1000 (AM preset, 12 kHz):
- Little-endian: range -31726 to 32527 (noise - full int16 range)
- Big-endian:    range 2669 to 5358 (clean audio with DC offset)

Listening test: big-endian audio is PERFECT WWV broadcast.

THE FIX:
--------
Changed from:
  np.frombuffer(payload, dtype=np.int16)   # Little-endian default

To:
  np.frombuffer(payload, dtype='>i2')      # Big-endian (network order)

IMPACT:
-------
Fixes BOTH:
1. audio_streamer.py - Now produces clean demodulated audio
2. grape_rtp_recorder.py - Now records correct IQ for science data

This affects ALL data recorded since the beginning - all IQ samples
were byte-swapped. Any doppler measurements from old data are INVALID.

TESTING:
--------
- Captured 10 seconds of SSRC 1000 (AM, 12kHz) with correct byte order
- Audio is clean WWV: voice, 1000 Hz tone, time ticks all audible
- SNR matches radiod report (39 dB)

+++ src/signal_recorder/grape_rtp_recorder.py
@@ -919,8 +919,9 @@ class GRAPEChannelRecorder:
             return
             
         # Unpack as interleaved I/Q int16, then normalize to float
+        # CRITICAL: RTP payloads use network byte order (BIG-ENDIAN)
         try:
-            samples_int16 = np.frombuffer(payload, dtype=np.int16).reshape(-1, 2)
+            samples_int16 = np.frombuffer(payload, dtype='>i2').reshape(-1, 2)  # Big-endian!
             # Normalize int16 (-32768 to 32767) to float (-1.0 to 1.0)
             samples = samples_int16.astype(np.float32) / 32768.0
             iq_samples = samples[:, 0] + 1j * samples[:, 1]
